requirements.txt
Flask
waitress
gpiozero
python-vlc
st7789
Pillow
rpi-lgpio
spidev
Flask-SocketIO
plexapi
python-dotenv